requirements.txt
# Liquidline Bank Reconciliation Automation
# Dependencies

# Web Framework
streamlit>=1.29.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

# AI/ML
anthropic>=0.18.0
rapidfuzz>=3.5.0

# Database
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Excel Output
xlsxwriter>=3.1.0

# Date handling
python-dateutil>=2.8.0
